move the unit test from rail_base to rail_astro_tools
