+++ tests/creation/test_degraders.py
@@ -1,0 +1,84 @@
+import os
+from typing import Type
+
+import numpy as np
+import pandas as pd
+import pytest
+
+from rail.core.data import DATA_STORE, TableHandle
+from rail.core.util_stages import ColumnMapper
+from rail.creation.degradation.quantityCut import QuantityCut
+# from rail.creation.degradation.spectroscopic_selections import *
+
+
+@pytest.fixture
+def data():
+    """Some dummy data to use below."""
+
+    DS = DATA_STORE()
+    DS.__class__.allow_overwrite = True
+
+    # generate random normal data
+    rng = np.random.default_rng(0)
+    x = rng.normal(loc=26, scale=1, size=(100, 7))
+
+    # replace redshifts with reasonable values
+    x[:, 0] = np.linspace(0, 2, x.shape[0])
+
+    # return data in handle wrapping a pandas DataFrame
+    df = pd.DataFrame(x, columns=["redshift", "u", "g", "r", "i", "z", "y"])
+    return DS.add_data("data", df, TableHandle, path="dummy.pd")
+
+
+@pytest.fixture
+def data_forspec():
+    """Some dummy data to use below."""
+
+    DS = DATA_STORE()
+    DS.__class__.allow_overwrite = True
+
+    # generate random normal data
+    rng = np.random.default_rng(0)
+    x = rng.normal(loc=26, scale=1, size=(200000, 7))
+
+    # replace redshifts with reasonable values
+    x[:, 0] = np.linspace(0, 2, x.shape[0])
+
+    # return data in handle wrapping a pandas DataFrame
+    df = pd.DataFrame(x, columns=["redshift", "u", "g", "r", "i", "z", "y"])
+    return DS.add_data("data_forspec", df, TableHandle, path="dummy_forspec.pd")
+
+
+@pytest.mark.parametrize(
+    "cuts,error",
+    [
+        (1, TypeError),
+        ({"u": "cut"}, TypeError),
+        ({"u": dict()}, TypeError),
+        ({"u": [1, 2, 3]}, ValueError),
+        ({"u": [1, "max"]}, TypeError),
+        ({"u": [2, 1]}, ValueError),
+        ({"u": TypeError}, TypeError),
+    ],
+)
+def test_QuantityCut_bad_params(cuts, error):
+    """Test bad parameters that should return Type and Value errors"""
+    with pytest.raises(error):
+        QuantityCut.make_stage(cuts=cuts)
+
+
+def test_QuantityCut_returns_correct_shape(data):
+    """Make sure QuantityCut is returning the correct shape"""
+
+    cuts = {
+        "u": 0,
+        "y": (1, 2),
+    }
+    degrader = QuantityCut.make_stage(cuts=cuts)
+
+    degraded_data = degrader(data).data
+
+    assert degraded_data.shape == data.data.query("u < 0 & y > 1 & y < 2").shape
+    os.remove(degrader.get_output(degrader.get_aliased_tag("output"), final_name=True))
+
+
